# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Raffle Draw",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/customer/update/profile": {
            "patch": {
                "tags": [
                    "Mobile"
                ],
                "operationId": "596b063491107ec3ba1b1e65ad234cee",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "first_name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "last_name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pin Validate successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/get-flw-pub-key": {
            "get": {
                "tags": [
                    "General"
                ],
                "operationId": "6d5428dcdc447aedfd488cbc56ac60cc",
                "responses": {
                    "200": {
                        "description": "Pin Validate successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/get-flw-sec-key": {
            "get": {
                "tags": [
                    "General"
                ],
                "operationId": "a3f032b14cb27c29c0c9eabde6c41cc7",
                "responses": {
                    "200": {
                        "description": "Pin Validate successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/get-flw-enc-key": {
            "get": {
                "tags": [
                    "General"
                ],
                "operationId": "8fabe5ec5a3e09abb02b876fad69e03b",
                "responses": {
                    "200": {
                        "description": "Pin Validate successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/validate-pin": {
            "post": {
                "tags": [
                    "General"
                ],
                "operationId": "784c87e3d098959a8c0b53890e8ba1cd",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "query",
                        "description": "pin",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pin Validate successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/customer/forgot-password": {
            "post": {
                "tags": [
                    "Mobile"
                ],
                "operationId": "4774265050b5eba05aee884573d5b519",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}